requirements.txt
Pillow==10.3.0
psycopg2-binary==2.9.9
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Minify==0.50
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Script==2.0.6
gunicorn==20.0.4
SQLAlchemy==2.0.38
Werkzeug==2.3.7
Jinja2==3.1.6
WTForms==3.0.1
itsdangerous==2.2.0
MarkupSafe==3.0.3
bcrypt==5.0.0
email-validator==2.0.0
python-dotenv==1.0.0
sentry-sdk==2.41.0
